backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-decouple
Pillow
dj-database-url
psycopg2-binary
gunicorn